WIP: Migrate eject engine
