+++ src/middleware/electron.js
@@ -1,15 +1,11 @@
-import { clipboard, remote } from "electron";
+import { clipboard } from "electron";
 import uniq from "lodash/uniq";
-import Path from "path";
-import { statSync } from "fs-extra";
 import {
   REMOVE_CUSTOM_EVENT,
-  EJECT_ENGINE,
   PASTE_CUSTOM_EVENTS
 } from "../actions/actionTypes";
 import confirmDeleteCustomEvent from "../lib/electron/dialog/confirmDeleteCustomEvent";
 import confirmReplaceCustomEvent from "../lib/electron/dialog/confirmReplaceCustomEvent";
-import confirmEjectEngineDialog from "../lib/electron/dialog/confirmEjectEngineDialog";
 import {
   getScenes,
   getScenesLookup,
@@ -22,8 +18,6 @@ import { walkEvents, walkSceneSpecificEvents, walkActorEvents, filterEvents } fr
 import { EVENT_CALL_CUSTOM_EVENT } from "../lib/compiler/eventTypes";
 import { editScene, editActor, editTrigger, editCustomEvent } from "../actions";
 import l10n from "../lib/helpers/l10n";
-import ejectEngineToDir from "../lib/project/ejectEngineToDir";
-import confirmEjectEngineReplaceDialog from "../lib/electron/dialog/confirmEjectEngineReplaceDialog";
 
 export default store => next => action => {
   if (action.type === REMOVE_CUSTOM_EVENT) {
@@ -160,35 +154,6 @@ export default store => next => action => {
         );
       });
     }
-  } else if (action.type === EJECT_ENGINE) {
-    const cancel = confirmEjectEngineDialog();
-
-    if (cancel) {
-      return;
-    }
-
-    const state = store.getState();
-    const outputDir = Path.join(state.document.root, "assets", "engine");
-
-    let ejectedEngineExists;
-    try {
-      statSync(outputDir);
-      ejectedEngineExists = true;
-    } catch (e) {
-      ejectedEngineExists = false;
-    }
-
-    if (ejectedEngineExists) {
-      const cancel2 = confirmEjectEngineReplaceDialog();
-      if (cancel2) {
-        return;
-      }
-    }
-
-    ejectEngineToDir(outputDir).then(() => {
-      remote.shell.openItem(outputDir);
-    });
-
   } else if (action.type === PASTE_CUSTOM_EVENTS) {
 
     try {
